Refactor into separate files. RpcHttpServer WIP.
